requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-SocketIO==4.0.0
get==2019.4.13
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
numpy==1.16.3
post==2019.4.13
public==2019.4.13
pygame==1.9.6
python-engineio==3.5.2
python-socketio==4.0.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
typing-extensions==3.7.2
urllib3==1.25.2
Werkzeug==0.15.4
yarl==1.3.0
